% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consisScore.R
\name{consisScore}
\alias{consisScore}
\title{Consistency Score}
\usage{
consisScore(quest.pars, no.b, no.traits, all.items, quest, m.theta, d.bi)
}
\arguments{
\item{quest.pars}{TIRT parameters of the questionnaire (e.g., Mplus output)}

\item{no.b}{number of blocks in the questionnaire}

\item{no.traits}{number of traits measured by the questionnaire}

\item{all.items}{item coding table (item-trait correspondence, coding of items, ...)}

\item{quest}{which questionnaire "BT"/"SD"/"BI"/"OS"... (name in all.items)}

\item{m.theta}{thetas (estimates on traits)}

\item{d.bi}{data frame with binary outcomes (same length as m.fscores, hence sometimes a subsample)}
}
\value{
consistency score: vector (length = observations)
}
\description{
Ratio of the frequency of items with a probability higher than chance level
of being preferred to the number of items (Brown & Bartram, 2011).
}
\details{
Function depends on: load2mat(), designMatrix(), designLoadings().
}
