% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longOrder.R
\name{longOrder}
\alias{longOrder}
\title{longOrder}
\usage{
longOrder(IDs, d.b.o, no.b)
}
\arguments{
\item{IDs}{vector with ID variables}

\item{d.b.o}{data frame with ordered rank orders}

\item{no.b}{number of blocks}
}
\value{
matrix rows = observations, columns = rank orders,
cells = counts how often a rank order (col) was repeated by the participant (row)
}
\description{
Determine longest rank order for every observation and each item order (e.g., 123, 321, ...).
}
