% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longOrderAvg.R
\name{longOrderAvg}
\alias{longOrderAvg}
\title{longOrderAvg}
\usage{
longOrderAvg(m.lcm)
}
\arguments{
\item{m.lcm}{matrix with long combinations (from longOrder())}
}
\description{
Determine average over all six longOrder values for every person.
}
