% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longOrderMax.R
\name{longOrderMax}
\alias{longOrderMax}
\title{longOrderMax}
\usage{
longOrderMax(m.lcm)
}
\arguments{
\item{m.lcm}{matrix with long rank orders (from longOrder())}
}
\description{
Determine longest rank order for every observation.
Which rank order (e.g., 123, 321, ...) was repeated most often?
}
