% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseTimeIndex.R
\name{responseTimeIndex}
\alias{responseTimeIndex}
\title{Response Time Index}
\usage{
responseTimeIndex(df.time.vars, t_upper, t_lower)
}
\arguments{
\item{df.time.vars}{data frame with time measures per page or triplet}

\item{t_upper}{upper threshold}

\item{t_lower}{lower threshold}
}
\value{
data frame with binary coded data
}
\description{
To calculate the RTI, first, the coded response times C_{pt} are obtained from the response time of
each participant p on every triplet page t of the questionnaire.
Second, the RTI for each participant is determined as:
RTI_p=\frac{\sum_{t=1}^{T}C_{pt}}{k}
where k is the number of values of C_{pt} without the ones coded as missing (NA)
and T is the number of triplets in the questionnaire.
}
