% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameOrder.R
\name{sameOrder}
\alias{sameOrder}
\title{sameOrder}
\usage{
sameOrder(d.b.o, no.b)
}
\arguments{
\item{d.b.o}{data frame with ordered combinations}

\item{no.b}{number of blocks}
}
\description{
Determine proportion of triplets for which a person copied the presented order.
}
