% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortOrders.R
\name{sortOrders}
\alias{sortOrders}
\title{sortOrders}
\usage{
sortOrders(no.b, no.m, d.r, d.k)
}
\arguments{
\item{no.b}{number of blocks in the questionnaire of interest}

\item{no.m}{number of items per block (for triplets no.m = 3)}

\item{d.r}{data frame with raw data (unrecoded answers to triplets)}

\item{d.k}{data frame with key variables (order of triplets)}
}
\value{
data frame with rank orders of all blocks, rows = persons, columns = blocks
}
\description{
sort the rank orders of the raw answers to the randomized item blocks
}
\examples{
bl.com <- sortOrders(no.b = 20, no.m = 3, d.r = rawdata, d.k = keydata)
}
