% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripletVariance.R
\name{tripletVariance}
\alias{tripletVariance}
\title{tripletVariance}
\usage{
tripletVariance(d.b.o)
}
\arguments{
\item{d.b.o}{data frame with ordered rank orders (resulting from sortOrders())}
}
\value{
vector (length = observations) with triplet variance for each participant
}
\description{
Variance of a person’s ranking patterns across all triplets.
(Relativer Informationsgehalt; Eid,  Gollwitzer, & Schmitt, 2015)
H = -(1/ln(k)*sum(h_j *ln(h_j)))
k = number of possible combinations (in case of triplets 6)
h_j = proportion
}
